v1

+++ app/src/main/java/com/example/project2_atlas/Model/Nutrition.kt
@@ -1,9 +1,9 @@
 package com.example.project2_atlas.Model
 
 data class Nutrition (
-    val carbohydrates : Double,
-    val protein : Double,
-    val fat : Double,
-    val calories :Double,
-    val sugar:Double
+    val carbohydrates : Int,
+    val protein : Int,
+    val fat : Int,
+    val calories :Int,
+    val sugar:Int
 )
